requirements.txt
# Tune Agent Builder V3.0 - Production Requirements

# Core AI
anthropic>=0.18.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Clients
httpx>=0.26.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
aiosqlite>=0.19.0
# For PostgreSQL production:
# psycopg2-binary>=2.9.9

# Async Support
asyncio

# Data Processing
python-dateutil>=2.8.2

# Structured Logging
structlog>=24.1.0
python-json-logger>=2.0.7

# Error Handling & Reliability
tenacity>=8.2.3
circuitbreaker>=1.4.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Caching
redis>=5.0.1
aioredis>=2.0.1

# Monitoring & Observability
prometheus-client>=0.19.0
sentry-sdk[fastapi]>=1.40.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Template Engine
jinja2>=3.1.2

# Web Research & Scraping
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
# google-search-results>=2.4.2  # SerpAPI client

# Statistical Analysis
scipy>=1.11.0
numpy>=1.26.0
statsmodels>=0.14.0

# Advanced Analytics
pandas>=2.1.0

# ML/Content Quality
scikit-learn>=1.3.0
textstat>=0.7.3

# Content Safety
better-profanity>=0.7.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
